Title: 학점 계산기, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240102/학점 계산기/credit-calculator.c
@@ -3,17 +3,17 @@
 int main() {
     // 여기에 코드를 작성해주세요.
     int n;
-    float sum=0;
+    double sum=0;
     scanf("%d\n",&n);
-    float score[n];
+    double score[n];
     for(int i=0; i<n; i++){
-        scanf("%f ",&score[i]);
+        scanf("%lf ",&score[i]);
     }
     for(int i=0; i<n; i++){
         sum+=score[i];
     }
-    float ave = sum/(float)n;
-    printf("%.1f\n",ave);
+    double ave = sum/(double)n;
+    printf("%.1lf\n",ave);
     if(ave>=4.0){
         printf("Perfect");
     } else if(ave >= 3.0){
